Update se_like_I_like_it.user.js
I just realized the "history" variable could conflict with the window.history object.  No undesirable behavior has resulted from this, but I moved it into key{} just to be safe.  I also moved the key{} object to the global scope, as it seemed unnecessary to redefine it with every keypress.